New DontCleanupResults config parameter
The new config parameter disable automatic deletion of results directory after a
test is over. The new flag can be set in wptdriver.ini and is turned off by
default.

+++ agent/wptdriver/webpagetest.cc
@@ -161,7 +161,12 @@ bool WebPagetest::GetTest(WptTestDriver& test) {
   _lockScreen = 0;
 
 
-  DeleteDirectory(test._directory, false);
+  if (_settings._dontCleanupResults) {
+    WptTrace(loglevel::kError, _T("[wptdriver] - Results directory not deleted and left behind"));
+  }
+  else {
+    DeleteDirectory(test._directory, false);
+  }
 
 
   // build the url for the request
